Moves firebase object to re-frame database and only creates one auth listener
The initialization is a re-frame dispatch event now.

The auth listener is only called if a firebase object is found in the
local database.

The auth listener is only added once, this is controlled with the
boolean from the local database which knows if a auth listener has
already been added.