Adds tongue localization library
Replaces "Login with Google" with a localized text from the
localization file with an example on substitution.

The same login localized text could now be paired with
Google or Facebook for example. Might be useful.

We have two functions to localize text, one with substitution
and one without.

The default language can be set in localization.cljs if a
key is not found with the locale being used.

The locale comes from db.cljs, where re-frame dispatch can change it.
An example of this are the flag images at the front-page when you are
logged out. You see the login text change.