Adds signing up with registering username and description
Possibility to add a user description when you sign up for the first
time. Doesn't appear on next login if you opt out on the first time.

Stuff can be refactored further to clean out firebase-methods to
their own files and such. Now stuff works as a POC.

Doesn't work on firebase-serve, but fighweeling works..