Adds cypress tests for signing up
Actually enables email authentication. No error handling for
if the email is already found from the authenticated list.

Fixing that is an issue to the prettification/refactor of the email-signup.
Which is another Trello task. This keeps code changes to a minimum while
adding the tests for basic sign-up caseso