Changes travis to work with the new database and directories
Travis uses their own cloud mysql database and connects to it with
sequelize with the information in config.json. Therefore NODE_ENV
variable is needed.

Cypress tests now uses the new user created in the sequelize seed
demouser.

Travis starts the ring server for the background and then the frontend
server and waits for while for it to start then runs cypress ui tests.
If all is cool, you get a green run.